refactor(builder): Move Any out of parser
