+++ clap_builder/src/parser/matches/mod.rs
@@ -1,9 +1,7 @@
-mod any_value;
 mod arg_matches;
 mod matched_arg;
 mod value_source;
 
-pub use any_value::AnyValueId;
 pub use arg_matches::IdsRef;
 pub use arg_matches::RawValues;
 pub use arg_matches::Values;
@@ -11,6 +9,5 @@ pub use arg_matches::ValuesRef;
 pub use arg_matches::{ArgMatches, Indices};
 pub use value_source::ValueSource;
 
-pub(crate) use any_value::AnyValue;
 pub(crate) use arg_matches::SubCommand;
 pub(crate) use matched_arg::MatchedArg;
