+++ clap_builder/src/parser/mod.rs
@@ -10,8 +10,6 @@ mod validator;
 pub(crate) mod features;
 
 pub(crate) use self::arg_matcher::ArgMatcher;
-pub(crate) use self::matches::AnyValue;
-pub(crate) use self::matches::AnyValueId;
 pub(crate) use self::matches::{MatchedArg, SubCommand};
 pub(crate) use self::parser::Identifier;
 pub(crate) use self::parser::PendingArg;
