+++ clap_builder/src/parser/parser.rs
@@ -13,9 +13,9 @@ use crate::error::Result as ClapResult;
 use crate::mkeymap::KeyType;
 use crate::output::Usage;
 use crate::parser::features::suggestions;
-use crate::parser::AnyValue;
 use crate::parser::{ArgMatcher, SubCommand};
 use crate::parser::{Validator, ValueSource};
+use crate::util::AnyValue;
 use crate::util::Id;
 use crate::ArgAction;
 use crate::INTERNAL_ERROR_MSG;
