+++ clap_builder/src/util/mod.rs
@@ -1,5 +1,6 @@
 #![allow(clippy::single_component_path_imports)]
 
+mod any_value;
 pub(crate) mod flat_map;
 pub(crate) mod flat_set;
 mod graph;
@@ -8,6 +9,8 @@ mod str_to_bool;
 
 pub use self::id::Id;
 
+pub(crate) use self::any_value::AnyValue;
+pub(crate) use self::any_value::AnyValueId;
 pub(crate) use self::flat_map::Entry;
 pub(crate) use self::flat_map::FlatMap;
 pub(crate) use self::flat_set::FlatSet;
